Rename count_code_lines to individual_metrics. Similar changes to test
