+++ src/print_table.py
@@ -1,7 +1,7 @@
 """This program takes the path to the repo and prints the data table."""
 
 from prettytable import PrettyTable
-import count_code_lines
+import individual_metrics
 import json_handler
 
 
@@ -43,12 +43,12 @@ def print_individual_in_table(file_name):
 if __name__ == "__main__":
     FILE_NAME = input("Enter the name of the file : ")
     # FILE_NAME = "contributor_data_template"
-    DATA = count_code_lines.calculate_individual_metrics(FILE_NAME)
+    DATA = individual_metrics.calculate_individual_metrics(FILE_NAME)
     if DATA == {}:
         REPO_PATH = input("Enter the path to the repo : ")
-        count_code_lines.add_raw_data_to_json(REPO_PATH, FILE_NAME)
+        individual_metrics.add_raw_data_to_json(REPO_PATH, FILE_NAME)
         print("processing data again")
-        DATA = count_code_lines.calculate_individual_metrics(FILE_NAME)
+        DATA = individual_metrics.calculate_individual_metrics(FILE_NAME)
     print("Adding processed data to selected json file...")
     # Write reformatted dictionary to json
     json_handler.add_entry(DATA, FILE_NAME)
